backend/pyproject.toml
[project]
name = "voicelab-pro-backend"
version = "0.1.0"
description = "VoiceLab Pro Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "elevenlabs>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

